requirements.txt
gspread==6.1.2
pandas==2.2.2
google-auth==2.33.0
PyMuPDF==1.24.9
python-dotenv==1.0.1
